+++ src/jennifer/migration/table_builder/create_table.cr
@@ -77,7 +77,7 @@ module Jennifer
 
         # Adds a reference.
         #
-        # The reference column is an `integer` by default, *type* argument can be used to specify a different type.
+        # The reference column is an `bigint` by default, *type* argument can be used to specify a different type.
         #
         # If *polymorphic* option is `true` - additional string field `"#{name}_type"` is created and foreign key is
         # not added.
@@ -87,10 +87,10 @@ module Jennifer
         #
         # ```
         # reference :user
-        # reference :order, :bigint
+        # reference :order, :integer
         # reference :taggable, {:polymorphic => true}
         # ```
-        def reference(name, type : Symbol = :integer, options : Hash(Symbol, AAllowedTypes) = DbOptions.new)
+        def reference(name, type : Symbol = :bigint, options : Hash(Symbol, AAllowedTypes) = DbOptions.new)
           column = Inflector.foreign_key(name)
           is_null = options.has_key?(:null) ? options[:null] : true
